requirements.txt
watchdog>=2.1.9
requests>=2.28.1
python-i18n>=0.3.9
pydantic>=1.9.0
pyserial>=3.5
cachetools>=5.0.0
python-dotenv>=0.19.2
structlog>=21.5.0
google-cloud-translate>=3.7.4
numpy>=1.24.0
pillow>=10.0.0